Ref: admin page refactored #129
