+++ src/app/admin/layout.js
@@ -1,0 +1,14 @@
+//import { Suspense } from "react";
+
+export const metadata = {
+    title: "Admin",
+    description: "Admin management and participation platform",
+};
+
+export default function AdminLayout({ children }) {
+    return (
+        <>
+            {children}
+        </>
+    );
+}
